module2-task1: some updates
