+++ src/mock/point.js
@@ -1,8 +1,13 @@
 import {getRandomElement, getRandomInteger, shuffleElements} from '../utils';
-import {DESTINATIONS, OFFER_TYPES, OFFERS} from '../const';
+import {OFFER_TYPES} from '../const';
+import {getDestinations} from './destinations';
+import {getOffers} from './offers';
+
+const initialDestinations = getDestinations();
+const initialOffers = getOffers();
 
 const getOfferIds = (type) => {
-  const offers = OFFERS.find((offer) => offer.type === type).offers;
+  const offers = initialOffers.find((offer) => offer.type === type).offers;
 
   return shuffleElements(offers).slice(0, getRandomInteger(0, offers.length)).map((offer) => offer.id);
 };
@@ -16,7 +21,7 @@ export const generatePoint = () => {
     basePrice: getRandomInteger(500, 1500),
     dateFrom: `2022-01-0${getRandomInteger(1, 4)}T11:55:56`,
     dateTo: `2022-01-0${getRandomInteger(4, 9)}T15:55:13`,
-    destination: getRandomElement(DESTINATIONS).id,
+    destination: getRandomElement(initialDestinations).id,
     isFavorite: Boolean(getRandomInteger(0, 1)),
     offers,
     type,
